listen for spring transaction as commandContextCloseListener doesn't know if spring transaction is ended
