--- activiti-cloud-services-runtime-bundle/activiti-cloud-services-connectors/src/main/java/org/activiti/cloud/services/events/listeners/IntegrationEventsAggregator.java
@@ -1,50 +1,0 @@
-/*
- * Copyright 2017 Alfresco, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.activiti.cloud.services.events.listeners;
-
-import org.activiti.services.connectors.behavior.IntegrationProducerCommandContextCloseListener;
-import org.activiti.services.connectors.model.IntegrationRequestEvent;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.messaging.Message;
-import org.springframework.stereotype.Component;
-
-@Component
-public class IntegrationEventsAggregator extends BaseCommandContextEventsAggregator<Message<IntegrationRequestEvent>, IntegrationProducerCommandContextCloseListener> {
-
-    private final IntegrationProducerCommandContextCloseListener closeListener;
-
-    @Autowired
-    public IntegrationEventsAggregator(IntegrationProducerCommandContextCloseListener closeListener) {
-        this.closeListener = closeListener;
-    }
-
-    @Override
-    protected Class<IntegrationProducerCommandContextCloseListener> getCloseListenerClass() {
-        return IntegrationProducerCommandContextCloseListener.class;
-    }
-
-    @Override
-    protected IntegrationProducerCommandContextCloseListener getCloseListener() {
-        return closeListener;
-    }
-
-    @Override
-    protected String getAttributeKey() {
-        return IntegrationProducerCommandContextCloseListener.PROCESS_ENGINE_INTEGRATION_EVENTS;
-    }
-
-}
