+++ activiti-cloud-services-runtime-bundle/activiti-cloud-services-connectors/src/main/java/org/activiti/services/connectors/behavior/MQServiceTaskBehavior.java
@@ -18,59 +18,59 @@ package org.activiti.services.connectors.behavior;
 
 import java.util.Date;
 
-import org.activiti.bpmn.model.ServiceTask;
-import org.activiti.cloud.services.events.listeners.IntegrationEventsAggregator;
+import org.activiti.cloud.services.events.ProcessEngineChannels;
+import org.activiti.cloud.services.events.configuration.RuntimeBundleProperties;
 import org.activiti.engine.delegate.DelegateExecution;
 import org.activiti.engine.impl.bpmn.behavior.AbstractBpmnActivityBehavior;
 import org.activiti.engine.impl.delegate.TriggerableActivityBehavior;
 import org.activiti.engine.impl.persistence.entity.integration.IntegrationContextEntity;
 import org.activiti.engine.impl.persistence.entity.integration.IntegrationContextManager;
-import org.activiti.services.connectors.model.IntegrationRequestEvent;
+import org.activiti.services.connectors.IntegrationRequestSender;
+import org.activiti.services.connectors.channel.ProcessEngineIntegrationChannels;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.messaging.Message;
-import org.springframework.messaging.support.MessageBuilder;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.support.TransactionSynchronizationManager;
 
 @Component
 public class MQServiceTaskBehavior extends AbstractBpmnActivityBehavior implements TriggerableActivityBehavior {
 
-    private static final String CONNECTOR_TYPE = "connectorType";
     private final IntegrationContextManager integrationContextManager;
-    private final IntegrationEventsAggregator eventsAggregator;
+    private final ProcessEngineIntegrationChannels integrationChannels;
+    private static final Logger LOGGER = LoggerFactory.getLogger(IntegrationRequestSender.class);
+    private final RuntimeBundleProperties runtimeBundleProperties;
+    private final ProcessEngineChannels processEngineChannels;
+
 
     @Autowired
     public MQServiceTaskBehavior(IntegrationContextManager integrationContextManager,
-                                 IntegrationEventsAggregator eventsAggregator) {
+                                 ProcessEngineIntegrationChannels processEngineIntegrationChannels,
+                                 RuntimeBundleProperties runtimeBundleProperties,
+                                 ProcessEngineChannels processEngineChannels) {
         this.integrationContextManager = integrationContextManager;
-        this.eventsAggregator = eventsAggregator;
+        this.integrationChannels = processEngineIntegrationChannels;
+        this.runtimeBundleProperties = runtimeBundleProperties;
+        this.processEngineChannels = processEngineChannels;
     }
 
     @Override
     public void execute(DelegateExecution execution) {
         IntegrationContextEntity integrationContext = storeIntegrationContext(execution);
-        eventsAggregator.add(buildMessage(execution,
-                                          integrationContext));
+
+        IntegrationRequestSender integrationRequestSender = new IntegrationRequestSender(integrationContextManager, integrationChannels, runtimeBundleProperties, processEngineChannels, integrationContext, execution);
+        registerTransactionSynchronization(integrationRequestSender);
     }
+
+    protected void registerTransactionSynchronization(IntegrationRequestSender integrationRequestSender) {
+        TransactionSynchronizationManager.registerSynchronization(integrationRequestSender);
+    }
+
 
     private IntegrationContextEntity storeIntegrationContext(DelegateExecution execution) {
         IntegrationContextEntity integrationContext = buildIntegrationContext(execution);
         integrationContextManager.insert(integrationContext);
         return integrationContext;
-    }
-
-    private Message<IntegrationRequestEvent> buildMessage(DelegateExecution execution,
-                                                          IntegrationContextEntity integrationContext) {
-        IntegrationRequestEvent event = new IntegrationRequestEvent(execution.getProcessInstanceId(),
-                                                                    execution.getProcessDefinitionId(),
-                                                                    integrationContext.getExecutionId(),
-                                                                    integrationContext.getId(),
-                                                                    execution.getVariables());
-
-        String implementation = ((ServiceTask) execution.getCurrentFlowElement()).getImplementation();
-        return MessageBuilder.withPayload(event)
-                .setHeader(CONNECTOR_TYPE,
-                           implementation)
-                .build();
     }
 
     private IntegrationContextEntity buildIntegrationContext(DelegateExecution execution) {
